Issue 897: renamed LM validation constraint once more.
